+++ src/main.js
@@ -1,10 +1,10 @@
 import log from './utils/log.js'
-import * as imageAnchor from './aspects/image-anchor.js'
+import * as imageOffset from './aspects/image-offset.js'
 import * as instantTokenPreview from './aspects/instant-token-preview.js'
 
 Hooks.on('ready', function()
 {
-    imageAnchor.activate()
+    imageOffset.activate()
     instantTokenPreview.activate()
     log('Ready')
 })
